expense-tracker/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pydantic[email]==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.4
anthropic==0.28.0
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
